Refactoring class name and package
